refactor deferred logic
